Revert health and damage amounts to float values from int, fixes issue with health always showing 0%.

+++ Assets/_Scripts/Entities/Enemies/BossEnemy.cs
@@ -117,7 +117,7 @@ public class BossEnemy : ShieldedEnemy
     /// <summary>Checks if the entity can be attacked, and attacks them if so.</summary>
     /// <param name="target">The entity to attack.</param>
     /// <param name="damage">The damage to deal to the entity.</param>
-    public override void Attack(Entity target, int damage)
+    public override void Attack(Entity target, float damage)
     {
         base.Attack(target, damage + 5 * Mathf.RoundToInt(bossScaleMult));
     }
